refactoring and updating to match guide
