+++ examples/tutorial/reactivex/shared.py
