+++ examples/tutorial/step3/shared.py
