+++ examples/tutorial/step5/shared.py
