+++ examples/tutorial/step6/shared.py
