+++ examples/tutorial/step7/shared.py
