+++ examples/tutorial/step8/shared.py
